recognition/Siamese_47425626/requirements.txt
certifi==2024.8.30
contourpy==1.3.0
cycler==0.12.1
exchange_calendars==4.5.6
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
korean-lunar-calendar==0.3.1
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
packaging==24.1
pandas==2.2.3
pandas_market_calendars==4.4.1
pillow==11.0.0
polygon-api-client==1.14.2
pyluach==2.2.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-metric-learning==2.6.0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sympy==1.13.1
threadpoolctl==3.5.0
toolz==1.0.0
torch==2.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websockets==12.0
